+++ resolwe/flow/models/utils.py
@@ -159,6 +159,12 @@ def validate_schema(instance, schema, test_required=True, path_prefix=None):
             except jsonschema.exceptions.ValidationError as ex:
                 raise ValidationError(ex.message)
 
+            choices = _schema.get('choices', [])
+            allow_custom_choice = _schema.get('allow_custom_choice', False)
+            if choices and not allow_custom_choice and field not in choices:
+                raise ValidationError(
+                    "Value must match one of predefined choices.")
+
             if type_ == 'basic:file:':
                 validate_file(field, _schema.get('validate_regex'))