+++ docs/CHANGELOG.rst
@@ -33,6 +33,8 @@ Added
 - Support ordering by version for descriptor schema
 - Support process test case based on Resolwe's ``TransactionTestCase``
 - Add Null executor
+- If ``choices`` are defined in JSON schema, value of field is 
+  validated with them
 
 Changed
 -------